# frontend/src/api/openapi-snapshot.json
{
  "components": {
    "schemas": {
      "AccountDeletion": {
        "description": "Confirmation d'une fermeture de compte.\n\nLe mot de passe est exigé pour la même raison qu'au changement : un jeton volé ne doit pas\nsuffire à détruire les chantiers d'une entreprise.",
        "properties": {
          "current_password": {
            "maxLength": 128,
            "title": "Current Password",
            "type": "string"
          }
        },
        "required": [
          "current_password"
        ],
        "title": "AccountDeletion",
        "type": "object"
      },
...
      "DemoProjectCreated": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "title": "Project Id",
            "type": "integer"
          }
        },
        "required": [
          "project_id",
          "name"
        ],
        "title": "DemoProjectCreated",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PasswordForgot": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "PasswordForgot",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Éditeur de plan de rénovation 2D → 3D",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/demo-project": {
      "post": {
        "description": "Sème le chantier de démonstration dans l'organisation par défaut du compte.\n\nRéservée à un espace **vierge** : dès qu'un chantier existe, la route répond 409 et n'écrit\nrien. C'est ce qui permet à l'artisan de supprimer la démonstration pour de bon — un objet\nqu'on ne peut pas jeter est plus irritant qu'un état vide — et c'est aussi le garde-fou contre\ndeux onglets qui la demanderaient en même temps.",
        "operationId": "create_demo_api_auth_demo_project_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DemoProjectCreated"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Demo",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/me": {
      "delete": {
        "description": "Ferme le compte et efface ses données (RGPD art. 17).\n\nRefusé — 409 — quand le compte est le dernier propriétaire accepté d'une organisation qui\ncompte d'autres membres. Ce n'est pas une restriction du droit à l'effacement : c'est le refus\nde détruire les données d'un tiers au passage. `project.owner_id` porte un\n`ON DELETE CASCADE` : partir emporterait tous les chantiers **créés** par ce compte, y compris\nceux que ses collègues éditent tous les jours. Le message nomme les organisations à transmettre\nd'abord.",
        "operationId": "delete_current_account_api_auth_me_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDeletion"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Current Account",
        "tags": [
          "auth"
        ]
      },
      "get": {
        "operationId": "read_current_user_api_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Current User",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me/export": {
      "get": {
        "description": "Export complet des données du compte, en JSON (RGPD art. 15 et 20).\n\nLe périmètre est exactement celui que les routes de l'API autorisent : les organisations dont\nle compte est membre **accepté**, et ce qu'elles portent. Un export plus large serait une\nfuite entre locataires déguisée en conformité, ce que `tests/test_rgpd.py` vérifie.",
        "operationId": "export_current_account_api_auth_me_export_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Export Current Account Api Auth Me Export Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Export Current Account",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/password": {
      "patch": {
        "description": "Change le mot de passe d'un compte connecté.\n\nLe mot de passe actuel est exigé : un jeton d'accès volé ne doit pas suffire à verrouiller le\ntitulaire hors de son propre compte. C'est la différence entre une session compromise, dont on\nreprend la main, et un compte perdu.\n\nToutes les sessions sont révoquées, **y compris celle-ci** — d'où la paire de jetons rendue en\nréponse. Sans elle, l'utilisateur serait déconnecté par son propre changement de mot de passe\net croirait à un échec.",
        "operationId": "change_password_api_auth_password_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Change Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/password/forgot": {
      "post": {
        "description": "Demande un lien de réinitialisation. Répond toujours 202, avec le même corps.\n\nLimité en débit comme la connexion et l'inscription : sans ça, la route serait le moyen le\nmoins cher d'inonder de messages une adresse qu'on n'a pas, et de périmer en boucle le lien\nqu'un utilisateur légitime vient de recevoir.",
        "operationId": "forgot_password_api_auth_password_forgot_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordForgot"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forgot Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/password/reset": {
      "post": {
        "description": "Consomme un jeton de réinitialisation et pose le nouveau mot de passe.\n\nUn jeton inconnu, expiré ou déjà utilisé rend le même 400 : distinguer ces cas dirait à qui\nprésente un jeton au hasard s'il a existé. Le refus n'est pas un oracle d'énumération pour\nautant — le secret est le jeton, pas l'adresse.\n\nLa réinitialisation révoque toutes les sessions : quelqu'un a peut-être pris la main sur le\ncompte, et c'est précisément le moment de le mettre dehors.",
        "operationId": "reset_password_api_auth_password_reset_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordReset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PasswordResetDone"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Password",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "description": "Échange un jeton de rafraîchissement contre une nouvelle paire (rotation).\n\nLe cookie est la source de vérité du navigateur, mais un jeton **explicitement** fourni dans\nle corps l'emporte : un client qui en présente un demande à valider celui-là, et se replier\nen silence sur le cookie ferait passer un jeton invalide pour bon.",
        "operationId": "refresh_api_auth_refresh_post",
        "parameters": [
          {
            "in": "cookie",
            "name": "refresh_token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Refresh Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RefreshRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/register": {
      "post": {
        "description": "Inscription.\n\nRépond toujours 202 avec le même corps, que l'adresse soit libre ou déjà prise (voir\n`RegistrationAccepted`). L'inscription est elle aussi limitée en débit : sans ça, elle\nfournirait une source inépuisable de comptes valides pour contourner la limitation de la\nconnexion.",
        "operationId": "register_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}